useSearchFetch 커스텀hook 작성
